Show footer social links on mobile
On screen sizes under a mobile breakpoint the social (twitter/github/discord) links were hidden. I replaced that styling with some styles to center the icons to match the centered Tadoku logo in the footer.

The "Built by antonve" link was given a similar treatment, but the only change was to make the margins to be symmetrical instead of the bottom margin being twice the top. This looks better in my opinion when everything is closer.