Added installing Bazel as a local dev setup step
Bazel is required to build things and needs to be installed, I added it to the docs.

I also rearranged the install instructions to mention installing Helm and Bazel before Tilt, as the Tilt setup is longer and more involved (depending on your OS). During install on another machine I forgot there was anything beyond Tilt and all of its dependencies and had to re-discover that Helm was needed. Putting the quicker to install dependencies first should hopefully prevent me or other people from forgetting them in the future.

I also removed the "according to  their installation instructions" part of the Tilt install step, it seemed redundant, but I can revert it if desired.