Renamed database to include the prefix `tadoku-`
The postgres operator refused to do anything without the `teamId` (`tadoku`) in the database cluster name. Adding this in the operator config name required changes to every other place that used that name so I had a lot of changes in a bunch of places.

The operator error message was:
```
time="2021-11-16T02:25:58Z" level=debug msg="skipping \"ADD\" event for the invalid cluster: name must match {TEAM}-{NAME} format" cluster-name=default/reading-contest-api-db pkg=controller
```

Without that I was getting lots of errors on the reading-contest-api and identity-api services saying it couldn't find postgres secrets, so that error seemed to be a blocker.

In the meantime (I'm not sure if this contributed to fixing the issue for sure) I changed the source for the postgres operator to use the [official chart](https://github.com/zalando/postgres-operator/blob/master/docs/quickstart.md#helm-chart) which at is using 1.7.1 instead of 1.6.2 like the old one was.

I also updated the postgres version used by the seed db container (it used 12, but the operator is configured to use 13).