Show update descriptions on mobile
The update description table has a bunch of columns that show or hide based on the viewport width, so the content can differ for mobile and desktop just with CSS. On mobile a single column shows where the content is described in a sentence instead, but it was missing the description and people during the last contest noticed they couldn't see what other people were reading.

I just added the description on a new line below the other update information, as to not crowd the update's sentence, as it'd look awkward to have an arbitrarily long description in there. And some people put their own extra information (like mentioning how far through a book/VN they are) in the description so it'd be extra weird in there. So on its own it is!

I also happened to notice that the sentence format for the update wasn't formatting the update day in UTC time, so at mobile breakpoints it was showing "Nov 16th" but at desktop breakpoints it was showing "Nov 17th", so I fixed that too.

I branched this off of the postgres database renaming changes, as I wasn't able to run Tadoku without that.