Easier to understand contest edit form
The first time I made a round I was a little confused by what "Description" and "Open" meant, so this change aims to make that more obvious.

I added a placeholder on "Description" to make it obvious it's the name of a contest, and as a hint for the format it should take. I also changed the "Open" label to say "Open to new participants" so it's obvious that it only controls whether new people can join the contest, and not whether people can submit updates. This keeps the "Open" keyword so a new admin can make a connection to what the "Open" tag thing means on the rounds table.