Show contest dates in UTC time on the contest management page
When viewing contests on the contest management page as an admin the dates were changed to the user's local timezone. When managing a contest that runs off of UTC this is really confusing to confirm if the contest was created for the correct date. This changes the date formatting function in use to show the start and end dates in UTC time.