Added a blog post for the 6th round of 2021
Deleted the oldest post and added a new one, as seems to be the way things go here.

One minor difference, because the first column in the language table is acting as a header I made them `<th scope=row>` instead, then made them styled the same way the other tables look (disabled the center alignment and bold default styling). I didn't bother doing this for the old posts as it'd be pretty tedious (but if you'd prefer I can do it), and just give it another year and the old posts will work their way out of the blog posts ;)