Added a modal to generate contest stats for a blog post
This adds a modal with a textarea that shows some blog post html for a contest with all the generic text that'd be needed.

The modal just does a bunch of stuff with document.createElement() then returns the innerHTML of a temporary div so we should avoid XSS if it somehow managed to get in there (I haven't checked the username validation rules, but I imagine with Japanese names I've seen it is probably fairly permissive).

This required a new endpoint for admin roles, and I used the queries you sent me but the response is still kind of messy and doesn't conform to the snake case style used by other http routes. This is still pretty WIP so I don't expect this to merge as-is.

I added an editorconfig file so nvim stops using tabs for TS files. I also added a package.json script that runs prettier because until I ran that I had a lot of code that didn't conform to prettier's style.