PR Feedback
This switches the manual dom manipulation to create the blog preview out in favor of react-dom's `renderToStaticMarkup`. This will render plain HTML without any of reacts's extra DOM attributes.

I also renamed `totalPages` to `totalAmount` to keep in line with the `amount` in use with the rankings, as well as maintaining the idea that Tadoku scores in points and not specifically pages which implies there's only one medium.

I also made the stats endpoint conform to the way of mapping data between casings like the other endpoints do. It now will send stats as json in snake case.

I also made the textarea a bit taller by default. You could really only see a line or two by default, not really useful for a textarea. Not that people are going to be doing any editing in this, but it's the thought that counts.